LE_Kotlin: moved Limitations.kt into new package for constants
