+++ Attt_LocalEngine_Kotlin/src/main/kotlin/constants/Limitations.kt
@@ -1,6 +1,6 @@
 @file:Suppress("unused")
 
-package elements
+package constants
 
 // all numbers & limitations are set here
