+++ Attt_LocalEngine_Kotlin/src/main/kotlin/logic/GameRules.kt
@@ -1,7 +1,7 @@
 package logic
 
-import elements.MAX_WINNING_LINE_LENGTH
-import elements.MIN_WINNING_LINE_LENGTH
+import constants.MAX_WINNING_LINE_LENGTH
+import constants.MIN_WINNING_LINE_LENGTH
 import publicApi.AtttPlayer
 
 /**
